Fixed Bad Interpreter Error
Fixed an issue that the last commit caused where the whole script file
wouldn't run and would give a bad interpreter error. The actual code
changes were made just so the underlying file could be fixed. I believe
a fat finger that changed the encoding was to blame for the error.