Fixed Adding PiAssist to Emulation Station
Fixed an issue where not all of the required resources were created in
order for the PiAssist to be added to Emulation Station. Also added a
message informing the user that the process was complete. This process
does not currently have any error handling, so it will always say it was
completed even if something didn't work.