Added ROM Scraper and Removed Old Code
A ROM scraper that was coded by SSELPH has been added to the PiAssist.
This can be accessed via the Miscellaneous menu option. Once the ROM
Scraper option has been chosen, it will ask the user which system they
would like to scrape and then proceeds to scrape the metadata for each
ROM in that system's folder.