Fixed WiFi Bug and Scraper Running while Emulation Station is Running
If you cancel out of the password entry screen for a WiFi network, it
will now actually cancel, instead of trying to connect with a blank
password. Also, since the scraper cannot be run while Emulation Station
is running, I've put in a check to let the user know they need to quit
Emulation Station and run the script from command line.