Fixed WiFi bug Introduced in Last Commit
The last commit had a bug that would manually enter in your WiFi SSID
only when "Other" was not selected, which was the exact opposite of what
it was supposed to do.  Along with fixing this, if the user cancels out
of selecting a SSID or manually entering one, they will be returned to
the WiFi menu instead of the process trying to continue on.